EES-5828 - removed shared Analytics service project in favour of each project that integrates with analytics having its own set of options and path resolvers
