--- src/GovUk.Education.ExploreEducationStatistics.Analytics.Services/Services/Interfaces/IAnalyticsPathResolver.cs
@@ -1,10 +1,0 @@
-namespace GovUk.Education.ExploreEducationStatistics.Analytics.Service.Services.Interfaces;
-
-public interface IAnalyticsPathResolver
-{
-    string PublicApiQueriesDirectoryPath();
-    
-    string PublicApiQueriesProcessingDirectoryPath();
-    
-    string PublicApiQueriesReportsDirectoryPath();
-}
