+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Api/Services/Interfaces/IAnalyticsPathResolver.cs
@@ -1,0 +1,6 @@
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Api.Services.Interfaces;
+
+public interface IAnalyticsPathResolver
+{
+    string PublicApiQueriesDirectoryPath();
+}
